Update invoices
Change the way invoices are created, user now taken to a case selection screen to choose a case to create an invoice for.
Updates to models, forms and urls required. Along with update to views.
Add in the pre-population of relevant fields from the case model when creating an invoice.
Add JavaScript to change the Due Date based on the term selected.